api/requirements.txt
kaggle
opencv-python
albumentations
pillow
pandas
numpy
fastapi
uvicorn
httpx
python-dotenv
requests
anthropic